move to separate file, for easier future deletion of old code
